Add BaseCleanupActivity and refactor cleanup screens
