+++ app/src/main/kotlin/com/d4rk/cleaner/app/clean/trash/ui/TrashActivity.kt
@@ -1,29 +1,13 @@
 package com.d4rk.cleaner.app.clean.trash.ui
 
 import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import com.d4rk.android.libs.apptoolkit.app.theme.style.AppTheme
+import androidx.compose.runtime.Composable
+import com.d4rk.cleaner.core.ui.BaseCleanupActivity
 
-class TrashActivity : AppCompatActivity() {
+class TrashActivity : BaseCleanupActivity() {
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-
-        setContent {
-            AppTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.background
-                ) {
-                    TrashScreen(activity = this@TrashActivity)
-                }
-            }
-        }
+    @Composable
+    override fun ScreenContent() {
+        TrashScreen(activity = this@TrashActivity)
     }
 }
