+++ app/src/main/kotlin/com/d4rk/cleaner/app/clean/whatsapp/details/ui/WhatsAppDetailsActivity.kt
@@ -1,19 +1,13 @@
 package com.d4rk.cleaner.app.clean.whatsapp.details.ui
 
 import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import com.d4rk.android.libs.apptoolkit.app.theme.style.AppTheme
+import androidx.compose.runtime.Composable
+import com.d4rk.cleaner.core.ui.BaseCleanupActivity
 import com.d4rk.cleaner.app.clean.whatsapp.summary.domain.actions.WhatsAppCleanerEvent
 import com.d4rk.cleaner.app.clean.whatsapp.summary.ui.WhatsappCleanerSummaryViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class WhatsAppDetailsActivity : AppCompatActivity() {
+class WhatsAppDetailsActivity : BaseCleanupActivity() {
 
     companion object {
         const val EXTRA_TYPE = "type"
@@ -21,26 +15,21 @@ class WhatsAppDetailsActivity : AppCompatActivity() {
 
     private val viewModel: WhatsappCleanerSummaryViewModel by viewModel()
     private val detailsViewModel: DetailsViewModel by viewModel()
+    private var type: String = ""
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        type = intent.getStringExtra(EXTRA_TYPE) ?: ""
         super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        val type = intent.getStringExtra(EXTRA_TYPE) ?: ""
-        setContent {
-            AppTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    DetailsScreen(
-                        viewModel = viewModel,
-                        detailsViewModel = detailsViewModel,
-                        title = type,
-                        onDelete = { viewModel.onEvent(WhatsAppCleanerEvent.DeleteSelected(it)) },
-                        activity = this
-                    )
-                }
-            }
-        }
+    }
+
+    @Composable
+    override fun ScreenContent() {
+        DetailsScreen(
+            viewModel = viewModel,
+            detailsViewModel = detailsViewModel,
+            title = type,
+            onDelete = { viewModel.onEvent(WhatsAppCleanerEvent.DeleteSelected(it)) },
+            activity = this
+        )
     }
 }
