+++ app/src/main/kotlin/com/d4rk/cleaner/app/images/picker/ui/ImagePickerActivity.kt
@@ -1,19 +1,14 @@
 package com.d4rk.cleaner.app.images.picker.ui
 
 import android.os.Bundle
-import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.result.PickVisualMediaRequest
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import com.d4rk.android.libs.apptoolkit.app.theme.style.AppTheme
+import androidx.compose.runtime.Composable
+import com.d4rk.cleaner.core.ui.BaseCleanupActivity
 
-class ImagePickerActivity : AppCompatActivity() {
+class ImagePickerActivity : BaseCleanupActivity() {
     private val viewModel: ImagePickerViewModel by viewModels()
 
     private val pickMediaLauncher =
@@ -23,17 +18,12 @@ class ImagePickerActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            AppTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.background
-                ) {
-                    ImagePickerComposable(activity = this@ImagePickerActivity, viewModel)
-                }
-            }
-        }
         selectImage()
+    }
+
+    @Composable
+    override fun ScreenContent() {
+        ImagePickerComposable(activity = this@ImagePickerActivity, viewModel)
     }
 
     fun selectImage() {
