+++ app/src/main/kotlin/com/d4rk/cleaner/core/ui/BaseCleanupActivity.kt
@@ -1,0 +1,39 @@
+package com.d4rk.cleaner.core.ui
+
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.d4rk.android.libs.apptoolkit.app.theme.style.AppTheme
+
+/**
+ * Activity that provides common Compose setup for cleanup screens.
+ */
+abstract class BaseCleanupActivity : AppCompatActivity() {
+
+    /**
+     * Compose content to be rendered by this activity.
+     */
+    @Composable
+    protected abstract fun ScreenContent()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            AppTheme {
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    ScreenContent()
+                }
+            }
+        }
+    }
+}
